added users and login

+++ app.py
@@ -1,60 +1,119 @@
 import os
-from flask import Flask, render_template, redirect, url_for, request, json
-from flask_sqlalchemy import SQLAlchemy
+from flask import Flask, flash, render_template, redirect, url_for, request, json
+
 from flask_migrate import Migrate
+from flask_seeder import FlaskSeeder
 from datetime import datetime
 import tzlocal
 import requests
+import flask_login
+from models import db, Channel, User
+from hashids import Hashids
+
+seeder = FlaskSeeder()
+hashids = Hashids()
 
 basedir = os.path.abspath(os.path.dirname(__file__))
 testDB = os.path.join(basedir, 'test.db')
 app = Flask(__name__)
+app.secret_key = 'ekDiAhArrhpwMRid3F6YEVaMk0d9U02Y'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL') or \
                                         'sqlite:///' + testDB
-db = SQLAlchemy(app)
+db.init_app(app)
 migrate = Migrate(app, db)
+seeder.init_app(app, db)
+login_manager = flask_login.LoginManager()
+login_manager.init_app(app)
 
 
-class Channel(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(80))
-    slug = db.Column(db.String(80))
-    slack_webhook = db.Column(db.String(300))
-    discord_webhook = db.Column(db.String(300))
-    count = db.Column(db.Integer)
-    user_id = db.Column(db.Integer)
-    active = db.Column(db.Boolean)
+class Auth(flask_login.UserMixin):
+    pass
 
 
-# class Issue(db.Model):
-#     id = db.Column(db.Integer, primary_key=True)
-#     title = db.Column(db.String(255))
-#     url = db.Column(db.String(255))
-#     location = db.Column(db.String(800))
-#     created_at = db.Column(db.Integer)
-#     channel_id = db.Column(db.Integer)
+@login_manager.user_loader
+def user_loader(email):
+    user = User.query.filter_by(email=email).first()
+
+    if not isinstance(user, User):
+        return
+
+    auth = Auth()
+    auth.id = user.email
+    auth.user = user
+    return auth
+
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    if request.method == 'GET':
+        return render_template("login.html", url=request.url_root)
+
+    email = request.form.get("email", None)
+    remember = True if not request.form.get("remember", None) else False
+
+    user = User.query.filter_by(email=email, password=request.form['password']).first()
+
+    if isinstance(user, User):
+        auth = Auth()
+        auth.id = user.email
+        auth.user = user
+
+        flask_login.login_user(auth, remember=remember)
+        return redirect(url_for('index'))
+
+    flash('Email or password incorrect. Please try again.', category='danger')
+    return redirect(url_for('index'))
+
+
+@app.route('/logout')
+def logout():
+    flask_login.logout_user()
+    return redirect(url_for('login'))
+
+
+@login_manager.unauthorized_handler
+def unauthorized_handler():
+    return redirect(url_for("login"))
 
 
 @app.route("/")
+@flask_login.login_required
 def index():
-    channels = Channel.query.all()
-    return render_template("index.html", channels=channels, url=request.url_root)
+    if flask_login.current_user.user.role == "admin":
+        channels = Channel.query.all()
+    else:
+        channels = Channel.query.filter_by(user_id=flask_login.current_user.user.id).all()
+    return render_template("project/index.html", channels=channels, url=request.url_root)
 
 
 @app.route("/detail/<string:id>")
+@flask_login.login_required
 def detail(id):
-    channel = Channel.query.filter_by(id=id).first()
-    return render_template("detail.html", channel=channel, url=request.url_root)
+    if flask_login.current_user.user.role == "admin":
+        channel = Channel.query.filter_by(id=id).first()
+    else:
+        channel = Channel.query.filter_by(id=id, user_id=flask_login.current_user.user.id).first()
+
+    if isinstance(channel, Channel):
+        return render_template("project/detail.html", channel=channel, url=request.url_root)
+    else:
+        flash("Project not found", "danger")
+        return redirect(url_for("index"))
 
 
 @app.route("/add", methods=["POST"])
+@flask_login.login_required
 def add_channel():
     name = request.form.get("name")
-    new_channel = Channel(name=name, slack_webhook="", discord_webhook="", count=0, active=True)
+
+    new_channel = Channel(name=name, slack_webhook="", user_id=flask_login.current_user.user.id, discord_webhook="",
+                          count=0, active=True)
     db.session.add(new_channel)
     db.session.commit()
-
+    new_channel.code = hashids.encode(new_channel.id + 100)
+    db.session.commit()
+    flash("Project added successfully", "success")
     return redirect(url_for("detail", id=new_channel.id))
 
 
@@ -64,7 +123,7 @@ def webhook(id):
     message = request.json.get("message")
     event = request.json.get("event")
     url = request.json.get("url")
-    channel = Channel.query.filter_by(id=id).first()
+    channel = Channel.query.filter_by(code=id).first()
     title = project_name + " - " + message if message else "Issue"
     if channel and project_name:
         channel.count = 1 if channel.count is None else (channel.count + 1)
@@ -127,6 +186,7 @@ def webhook(id):
 
 
 @app.route("/active/<string:id>")
+@flask_login.login_required
 def active_channel(id):
     channel = Channel.query.filter_by(id=id).first()
     channel.active = not channel.active
@@ -135,8 +195,17 @@ def active_channel(id):
 
 
 @app.route("/edit/<string:id>", methods=["POST"])
+@flask_login.login_required
 def edit(id):
-    channel = Channel.query.filter_by(id=id).first()
+    if flask_login.current_user.user.role == "admin":
+        channel = Channel.query.filter_by(id=id).first()
+    else:
+        channel = Channel.query.filter_by(id=id, user_id=flask_login.current_user.user.id).first()
+
+    if not isinstance(channel, Channel):
+        flash("Project not found", "danger")
+        return redirect(url_for("index"))
+
     name = request.form.get("name")
     discord_webhook = request.form.get("discord_webhook")
     slack_webhook = request.form.get("slack_webhook")
@@ -144,17 +213,30 @@ def edit(id):
     channel.discord_webhook = discord_webhook
     channel.slack_webhook = slack_webhook
     db.session.commit()
+    flash("Project update successfully", "success")
     return redirect(url_for("detail", id=channel.id))
 
 
 @app.route("/delete/<string:id>")
+@flask_login.login_required
 def delete_channel(id):
-    channel = Channel.query.filter_by(id=id).first()
+    if flask_login.current_user.user.role == "admin":
+        channel = Channel.query.filter_by(id=id).first()
+    else:
+        channel = Channel.query.filter_by(id=id, user_id=flask_login.current_user.user.id).first()
+
+    if not isinstance(channel, Channel):
+        flash("Project not found", "danger")
+        return redirect(url_for("index"))
+
     db.session.delete(channel)
     db.session.commit()
+    flash("Project deleted successfully", "success")
     return redirect(url_for("index"))
 
 
 if __name__ == "__main__":
-    db.create_all()
+    with app.app_context():
+        db.create_all()
+
     app.run(debug=True, host="0.0.0.0", port=8080)
